backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.12.4
sentence-transformers==2.7.0
chromadb==0.5.0
pypdf==4.0.1
google-generativeai==0.8.4
